refactor: Edge Functions テストヘルパーを test-helpers.ts に共通化
- withEnv / withMockFetch / jsonResponse を test-helpers.ts に抽出
- 既存 tmdb_test.ts / igdb_test.ts を含む各テストが同ヘルパーを参照
- Sonar の新規コード重複 (20.7%) を解消

また knip で検出された未使用 export (src/features/backlog/types.ts の GamePlatform / GameReleaseDates) を削除。Edge Functions 側 igdb.ts には独自のローカル型があり参照していなかった。

Refs #269
