+++ supabase/functions/_shared/test-helpers.ts
@@ -1,0 +1,64 @@
+export async function withEnv(
+  values: Record<string, string | undefined>,
+  run: () => Promise<void>,
+) {
+  const previous = new Map<string, string | undefined>();
+
+  for (const [key, value] of Object.entries(values)) {
+    previous.set(key, Deno.env.get(key));
+    if (value === undefined) {
+      Deno.env.delete(key);
+    } else {
+      Deno.env.set(key, value);
+    }
+  }
+
+  try {
+    await run();
+  } finally {
+    for (const [key, value] of previous.entries()) {
+      if (value === undefined) {
+        Deno.env.delete(key);
+      } else {
+        Deno.env.set(key, value);
+      }
+    }
+  }
+}
+
+function resolveFetchUrl(input: string | URL | Request): URL {
+  if (input instanceof Request) {
+    return new URL(input.url);
+  }
+
+  if (input instanceof URL) {
+    return input;
+  }
+
+  return new URL(input);
+}
+
+export async function withMockFetch(
+  handler: (url: URL, init?: RequestInit) => Response | Promise<Response>,
+  run: () => Promise<void>,
+) {
+  const originalFetch = globalThis.fetch;
+
+  globalThis.fetch = (async (input: string | URL | Request, init?: RequestInit) => {
+    const url = resolveFetchUrl(input);
+    return await handler(url, init);
+  }) as typeof fetch;
+
+  try {
+    await run();
+  } finally {
+    globalThis.fetch = originalFetch;
+  }
+}
+
+export function jsonResponse(payload: unknown, init: ResponseInit = {}) {
+  return new Response(JSON.stringify(payload), {
+    status: init.status ?? 200,
+    headers: { "Content-Type": "application/json", ...init.headers },
+  });
+}
